refactor: extract translate orchestration
